Add verbose compilation logging
This will have to be enabled via an env var, rather than statically in code.

+++ src/coreclr/interpreter/intops.h
@@ -22,7 +22,7 @@ extern const uint8_t g_interpOpLen[];
 extern const int g_interpOpDVars[];
 extern const int g_interpOpSVars[];
 extern const InterpOpArgType g_interpOpArgType[];
-extern const uint8_t* InterpNextOp(const uint8_t* ip);
+extern const int32_t* InterpNextOp(const int32_t* ip);
 
 // This, instead of an array of pointers, to optimize away a pointer and a relocation per string.
 extern const uint32_t g_interpOpNameOffsets[];
